requirements.txt
pytorch==1.7.1 
torchvision==0.8.2
torchaudio==0.7.2
opencv-python
pathlib